src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/primitive/ArrayPrimitiveLongTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayPrimitiveLongTextParser extends ArrayTextParser {

    public ArrayPrimitiveLongTextParser() {
        super(long.class, 0L);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Long.parseLong(element.getText());
    }

}
